refactor: move Highlight to its own file
